NFDIV-4887 Refactor
